Repo/package renamed, readme.md changes
