+++ README.md
@@ -13,6 +13,11 @@ There are currently two extensions currently implemented:
 
 **Important** Neither extension will override an existing header, it will only add a new header if none is already set.
 
+# FTHttp
+
+The fthttp package provides functions to create new `http.Client` with convenient default values for timing out and FT specific transport implementation. 
+
+
 # Usage
 
 Create a new `*http.Client` which sets a `User-Agent` of `PLATFORM-system-code/version` for all requests:
@@ -31,3 +36,7 @@ ctx := tidutils.TransactionAwareContext(context.TODO(), "tid_1234")
 req, err := http.NewRequest("GET", uri, nil)
 client.Do(req.WithContext(ctx))
 ```
+
+You can also directly use `fthttp.NewDefaultClient` or if you need a more customized version of it via `fthttp.NewClient(...)`.
+
+When you choose to use the default client constructor function, bare in mind that it depends on the env variables for setting the values on the client.
